Музыка и фон
Добавил музыку и фон, поправил игровое поле
